Improved high-level API & naming.
The term "parameter" was bothering me - I want to keep it for hyperparameters
(in case we want to automatically fit those). Also, the ability to mix
different types of observations seems like a bad idea, in the end. Therefore, I
decided to create model classes that are tied to particular types of
observations.