Marginal likelihood computation.
This commit improves the numerical stability of the batch fitter and makes it
possible to compute the (log-)marginal likelihood of a KickScore model (using
both the batch & recursive fitters).

There is still some work left (e.g., handling edge cases), and the computations
need to be thoroughly tested, but this is a big step forward.