Lower-bound for the log-marginal likelihood.
If we use "reverse KL" variational inference (with CVI), it makes sense
to use a lower-bound to the log-marginal likelihood to guide
hyper-parameter selection. This diff is a step towards that. Note,
however, that the interface is likely going to change, as in the current
state it is not much more than a hack.