Add piecewise constant kernel.
This kernel enables to model the skill of sports teams across seasons or
other piecewise linear processes. Its implementation required an API
change across all kernels; specifically, the `transition` and
`noise_cov` methods are no longer just a function of the time
difference---they are now a function of the two timestamps.