Make model serializable using `pickle`.
It is now possible to pickle a large kickscore model (addresses issue #4 on
GitHub). Previously, pickling a model with many items and observations would
fail with a "maximum recursion depth" error.

To fix this, I had to remove the property `observations` of the `Item` class.
This is an API-breaking change, though I don't believe this property was widely
used.